fix aggregator : remove AsyncTask
