+++ app/src/main/java/ch/epfl/sdp/contamination/PositionAggregator.java
@@ -10,9 +10,8 @@ import java.util.Date;
  * Then, is rounds this average to the nearest meter (i.e. rounding latitude/longitude to ex. 5 decimals), and sends this location along with the last time to a DataSender.
  **/
 public interface PositionAggregator {
-    // TODO: Changed for debug, restore values 300000 & 10
-    int WINDOW_FOR_LOCATION_AGGREGATION = 5000; // [ms] This is the frequency with which the (mean) position will be uploaded. actual : 5min
-    int MAXIMAL_NUMBER_OF_LOCATIONS_PER_AGGREGATION = 1; // 100 locations per aggregation. Before release choose a big number
+    int WINDOW_FOR_LOCATION_AGGREGATION = 300000; // [ms] This is the frequency with which the (mean) position will be uploaded. actual : 5min
+    int MAXIMAL_NUMBER_OF_LOCATIONS_PER_AGGREGATION = 10;//WINDOW_FOR_LOCATION_AGGREGATION/10000; // default : 1 location every 10 seconds
     /**
      *
      * Adds a position to the position list. Every WINDOW_FOR_LOCATION_AGGREGATION time, it should send the mean value of the
